+++ samples/Arlecchino.Sample/Program.cs
@@ -37,6 +37,7 @@ builder.Services
     .AddGeneratedViews()
     .AddGeneratedStores()
     .AddGeneratedCommands()
+    .UseTheme(ThemePalette.Arlecchino)
     .UseLatinOnlyInput()
     .UseMouse()
     .StartAt(ViewKind.Default);
@@ -84,6 +85,7 @@ static void Frame(string view, string size)
         .AddGeneratedStores()
         .AddCommand<AboutCommand>()
         .AddCommand<QuitCommand>()
+        .UseTheme(ThemePalette.Arlecchino)
         .WithoutHostedService();
 
     services.AddSingleton<IHostApplicationLifetime, NullLifetime>();